fix uses of prompter function to allow passing null pointer
Change the function prototypes that take a callback of type
``krb5_prompter_fct'' to use Option<krb5_prompter_fct>. This
affects two functions:

    - krb5_get_init_creds_password()
    - krb5_init_creds_init()

According to the MIT Kerberos docs the prompter callback is only
relevant if no password is supplied [0]. libkrb5 checks for NULL
and will only invoke the callback if it is present. The correct
representation in Rust is thus an Option<fn()> [1].

Without the Option wrapping, LLVM will treat null pointers passed
from Rust (by transmuting to the function type) as UB and
eliminate the function call altogether along with any code that
depends on it. This can be observed both in release and debug
builds.

[0] https://web.mit.edu/kerberos/krb5-latest/doc/appdev/refs/api/krb5_get_init_creds_password.html#c.krb5_get_init_creds_password
[1] https://doc.rust-lang.org/nomicon/ffi.html#the-nullable-pointer-optimization