make responder callback nullable
The responder callback is a generalized version of the older
prompter functionality [0]. Also there is only one user of the
prototype, ``krb5_get_init_creds_opt_set_responder()``, and it is
also the only way to remove an existing callback by passing a
NULL pointer. To allow this the callback interface should be
represented as an Option<fn()> on the Rust end.

In libkrb5 the invocation of the responder (preauth2.c) is
guarded by a NULL check so this is safe.

[0] https://k5wiki.kerberos.org/wiki/Projects/Responder