make checksum callback nullable
The checksum callback will only be used if present; otherwise
libkrb5 will default the builtin checksum handler. API wise it is
thus perfectly valid for the handler argument to be NULL and Rust
should use an Option<fn()> to reflect that.

libkrb5 checks for NULL before invoking the callback
(mk_req_ext.c) so this is safe.